+++ tests/ds/test_plots.py
@@ -1,6 +1,5 @@
 import os
 from datetime import datetime, timedelta
-from itertools import product
 from math import ceil
 from pathlib import Path
 
@@ -12,13 +11,10 @@ from pelutils.ds.plots import (
     Figure,
     base_colours,
     colours,
-    double_moving_avg,
-    exp_moving_avg,
     get_dateticks,
     histogram,
     linear_binning,
     log_binning,
-    moving_avg,
     normal_binning,
     tab_colours,
 )
@@ -62,102 +58,6 @@ def test_histogram():
 
     x, y = histogram(obs, ignore_zeros=True)
     assert (y > 0).all()
-
-
-class TestMovingAverage:
-    def test_moving_avg(self):
-        data = np.random.randn(100)  # noqa: NPY002
-        num_neighs = np.arange(1, 10)
-
-        # First, test with only y given
-        for n in num_neighs:
-            x, mov = moving_avg(data, neighbors=n)
-            assert len(x) == len(mov) == len(data) - 2 * n
-            assert mov.std() < data.std()
-
-        # Then test with both x and y given
-        for n in num_neighs:
-            x = np.linspace(0, 10, len(data))
-            x, mov = moving_avg(x, data, neighbors=n)
-            assert len(x) == len(mov) == len(data) - 2 * n
-            assert mov.std() < data.std()
-
-        # Test that it converges as expected
-        for n in num_neighs:
-            data = np.linspace(-10, 10, 1000)
-            dist_to_zero = np.linalg.norm(data - np.zeros_like(data))
-            prev_dist_to_zero = dist_to_zero
-            while len(data) > 2 * n:
-                _, data = moving_avg(data, neighbors=n)
-                dist_to_zero = np.linalg.norm(data - np.zeros_like(data))
-                assert 0 <= dist_to_zero < prev_dist_to_zero
-                prev_dist_to_zero = dist_to_zero
-
-    def test_exp_moving_avg(self):
-        data = np.random.randn(100)  # noqa: NPY002
-        # alpha is 1, so data should be unchanged
-        x, y = exp_moving_avg(data, alpha=1)
-        assert len(x) == len(y) == len(data)
-        assert np.all(np.isclose(data, y))
-
-        # alpha is 0, so data should be equal to the first point
-        x, y = exp_moving_avg(data, alpha=0)
-        assert len(x) == len(y) == len(data)
-        assert np.all(np.isclose(data[0], y))
-
-        # Test that reverse works. For alpha = 0, data should be
-        # equal to the last point
-        x, y = exp_moving_avg(data, alpha=0, reverse=True)
-        assert len(x) == len(y) == len(data)
-        assert np.all(np.isclose(data[-1], y))
-
-        # For monotically increasing data, the data should lag behind
-        # except for the first point, which is the same
-        data = np.linspace(-10, 10)
-        x, y = exp_moving_avg(data)
-        assert np.all(y[1:] < data[1:])
-        assert np.isclose(data[0], y[0])
-
-        # Corresponding test for reversed
-        data = np.linspace(-10, 10)
-        x, y = exp_moving_avg(data, reverse=True)
-        assert np.all(y[:-1] > data[:-1])
-        assert np.isclose(data[-1], y[-1])
-
-        # Test that it also works when supplying x
-        x = np.linspace(-np.pi, np.pi)
-        data = np.sin(x) + np.random.randn(len(x))  # noqa: NPY002
-        x_smooth, y_smooth = exp_moving_avg(x, data)
-        assert len(x) == len(data) == len(x_smooth) == len(y_smooth)
-        assert (x == x_smooth).all()
-
-    def test_double_moving_avg(self):
-        test_samples = np.arange(100, 500, step=65)
-        test_inner_neighbors = np.arange(1, 10)
-        test_outer_neighbors = np.arange(1, 10)
-        data = np.random.randn(200)  # noqa: NPY002
-
-        for s, i, o in product(test_samples, test_inner_neighbors, test_outer_neighbors):
-            x, y = double_moving_avg(data, inner_neighbors=i, outer_neighbors=o, samples=s)
-            assert len(x) == len(y) == s
-            assert y.std() < data.std()
-
-        # Test that it also works when supplying x
-        x = np.linspace(-np.pi, np.pi)
-        data = np.sin(x) + np.random.randn(len(x))  # noqa: NPY002
-        samples = 300
-        x_smooth, y_smooth = double_moving_avg(x, data, samples=300)
-        assert len(x_smooth) == len(y_smooth) == samples
-        assert y_smooth.std() < data.std()
-
-    def test_list_arg_avg(self):
-        # Test that it works when supplying lists instead of numpy arrays
-        moving_avg([1, 2, 3])
-        moving_avg([1, 2, 3], [1, 2, 3])
-        exp_moving_avg([1, 2, 3])
-        exp_moving_avg([1, 2, 3], [1, 2, 3])
-        double_moving_avg([1, 2, 3])
-        double_moving_avg([1, 2, 3], [1, 2, 3])
 
 
 class TestBinning: